+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2025-02-06
+
 ### Added
 
 - Async discover function `discover_async`
@@ -53,7 +55,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial public release
 
-[Unreleased]: https://github.com/lukasberbuer/weatherlink-live-local-python/compare/0.3.0...HEAD
+[Unreleased]: https://github.com/lukasberbuer/weatherlink-live-local-python/compare/0.4.0...HEAD
+[0.4.0]: https://github.com/lukasberbuer/weatherlink-live-local-python/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/lukasberbuer/weatherlink-live-local-python/compare/0.2.1...0.3.0
 [0.2.1]: https://github.com/lukasberbuer/weatherlink-live-local-python/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/lukasberbuer/weatherlink-live-local-python/compare/0.1.0...0.2.0